server/requirements.txt
Django==2.1.5
djangorestframework==3.9.1
djangorestframework-jwt==1.11.0
psycopg2==2.7.6.1
